requirements.txt
alembic==1.6.2
aniso8601==9.0.1
attrs==21.2.0
boto3==1.17.73
botocore==1.20.73
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cognitojwt==1.3.0
ecdsa==0.14.1
flasgger==0.9.5
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Cognito==1.17
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
mistune==0.8.4
psycopg2==2.8.6
pyasn1==0.4.8
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
SQLAlchemy==1.4.14
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
